feat: (extract custodian uri)
The is_swissdatacustodian_ontology_uri function is used to check if the URI matches the base URL.
If the URI matches the base URL, the fragment identifier is extracted and used as the value_uri. The fragment_identifier is extracted from the custodianURI to create a link to a specific section within the HTML page.